pyproject.toml
[tool.poetry]
name = "robot-showcase"
version = "0.1.0"
description = ""
authors = ["alexanderschmitz <alexander.schmitz98@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
robotframework = "^7.2.2"
opencv-python = "^4.11.0.86"
pyautogui = "^0.9.54"
numpy = "^2.2.5"
pillow = "^11.2.1"
pywinauto = "^0.6.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
